refactor APIResource and add VirtualMachine resources
